refactor(journeys-ui): fold nine one-function TreeBlock helpers into block/action (c55)
Block-tree reading was fragmented across nine sibling one-function
modules. The eight tree-read helpers (transformer, searchBlocks,
getStepHeading, getStepTheme, getGoalDetails, isActionBlock,
filterActionBlocks, getTextResponseLabel) move into libs/block beside
getCardChild, which already followed this layout; the one action-shaped
helper (getNextStepSlug) moves into libs/action.

Each old directory keeps a one-line re-export shim, so every existing
import path — @core/journeys/ui/<helper> aliases and relative
../<helper> imports alike — resolves unchanged; no consumer edits.
block/index.ts and action/index.ts now also export the helpers as the
canonical home. Files moved with git mv; specs move alongside their
subjects.
