+++ libs/journeys/ui/src/libs/action/index.ts
@@ -1,1 +1,2 @@
 export { handleAction } from './action'
+export { getNextStepSlug } from './getNextStepSlug'
