+++ libs/journeys/ui/src/libs/block/filterActionBlocks.ts
@@ -1,9 +1,9 @@
-import { TreeBlock } from '../block'
 import {
   BlockFields_CardBlock as CardBlock,
   BlockFields_StepBlock as StepBlock
-} from '../block/__generated__/BlockFields'
-import { ActionBlock, isActionBlock } from '../isActionBlock'
+} from './__generated__/BlockFields'
+import { ActionBlock, isActionBlock } from './isActionBlock'
+import type { TreeBlock } from './TreeBlock'
 
 export function filterActionBlocks(step?: TreeBlock<StepBlock>): ActionBlock[] {
   const card = step?.children[0] as TreeBlock<CardBlock> | undefined
