+++ libs/journeys/ui/src/libs/block/getGoalDetails.tsx
