+++ libs/journeys/ui/src/libs/block/getStepTheme.ts
@@ -2,8 +2,9 @@ import { ThemeMode, ThemeName } from '@core/shared/ui/themes'
 
 import { CardFields } from '../../components/Card/__generated__/CardFields'
 import { StepFields } from '../../components/Step/__generated__/StepFields'
-import type { TreeBlock } from '../block'
 import { JourneyFields } from '../JourneyProvider/__generated__/JourneyFields'
+
+import type { TreeBlock } from './TreeBlock'
 
 export function getStepTheme(
   block?: TreeBlock<StepFields> | TreeBlock<CardFields>,
