+++ libs/journeys/ui/src/libs/block/index.ts
@@ -7,4 +7,12 @@ export {
   treeBlocksVar,
   blockHistoryVar
 } from './block'
+export { filterActionBlocks } from './filterActionBlocks'
 export { getCardChild } from './getCardChild'
+export { getGoalDetails } from './getGoalDetails'
+export { getStepHeading, ORDERED_TYPOGRAPHY_VARIANTS } from './getStepHeading'
+export { getStepTheme } from './getStepTheme'
+export { getTextResponseLabel } from './getTextResponseLabel'
+export { isActionBlock, type ActionBlock } from './isActionBlock'
+export { searchBlocks } from './searchBlocks'
+export { transformer } from './transformer'
