+++ libs/journeys/ui/src/libs/block/transformer.spec.ts
@@ -1,4 +1,4 @@
-import { transformer } from '.'
+import { transformer } from './transformer'
 
 describe('transformer', () => {
   it('should change flat array into tree successfully', () => {
