+++ libs/journeys/ui/src/libs/block/transformer.ts
@@ -1,7 +1,7 @@
 import sortBy from 'lodash/sortBy'
 
-import type { TreeBlock } from '../block'
-import type { BlockFields } from '../block/__generated__/BlockFields'
+import type { BlockFields } from './__generated__/BlockFields'
+import type { TreeBlock } from './TreeBlock'
 
 export function transformer(data: BlockFields[]): TreeBlock[] {
   const tree: TreeBlock[] = []
