+++ libs/journeys/ui/src/libs/getGoalDetails/index.ts
@@ -1,1 +1,1 @@
-export { getGoalDetails } from './getGoalDetails'
+export { getGoalDetails } from '../block/getGoalDetails'
