+++ libs/journeys/ui/src/libs/getStepHeading/index.ts
@@ -1,1 +1,4 @@
-export { getStepHeading, ORDERED_TYPOGRAPHY_VARIANTS } from './getStepHeading'
+export {
+  getStepHeading,
+  ORDERED_TYPOGRAPHY_VARIANTS
+} from '../block/getStepHeading'
