+++ libs/journeys/ui/src/libs/getStepTheme/index.ts
@@ -1,1 +1,1 @@
-export { getStepTheme } from './getStepTheme'
+export { getStepTheme } from '../block/getStepTheme'
