+++ libs/journeys/ui/src/libs/getTextResponseLabel/index.ts
@@ -1,1 +1,1 @@
-export { getTextResponseLabel } from './getTextResponseLabel'
+export { getTextResponseLabel } from '../block/getTextResponseLabel'
