+++ libs/journeys/ui/src/libs/searchBlocks/index.ts
@@ -1,1 +1,1 @@
-export { searchBlocks } from './searchBlocks'
+export { searchBlocks } from '../block/searchBlocks'
