+++ libs/journeys/ui/src/libs/transformer/index.ts
@@ -1,1 +1,1 @@
-export { transformer } from './transformer'
+export { transformer } from '../block/transformer'
